Improved how points are submitted to the server
Removed duplicate upload AsyncTask(Fixes #2)
Added ProgressDialog to UploadActivity(when uploading multiple saved points)
Fixed skipping of first item in local DB when uploading(Fixes #12)
UploadActivity is now an invisible activity which does not appear in history or recents and will keep alive until progress dialog is gone(Fixes #10, Fixes #11)
Improved UploadData task to upload any number of points at a time, publishing each point to progress. It will cancel the task if it reaches a point that is not successful and not remove it from local storage.
Added some documentation and refactored some variables to have more obvious names(#8, #9)